refactor(Kotlin): Tidying to comply with Kotlin style guidelines
